Bob Ross Painting Dataset

This dataset contains detailed information about the paintings from Bob Ross's television show The Joy of Painting. Each row corresponds to one painting and includes metadata such as the season, episode, number of colors used, and links to the painting images and YouTube videos.

Usage

bob_ross

Format

A data frame with 403 rows and 26 variables:

img_src

Character. URL link to the image of the painting.

painting_title

Character. The title of the painting.

season

Numeric. The season number in which the painting was created.

episode

Numeric. The episode number within the season.

num_colors

Numeric. The total number of colors used in the painting.

youtube_src

Character. URL link to the YouTube video of the episode where the painting was made.

colors

Character. List of colors used in the painting.

color_hex

Character. List of hexadecimal values for the colors used in the painting.

Black_Gesso

Logical. Whether the color "Black Gesso" was used in the painting.

Bright_Red

Logical. Whether the color "Bright Red" was used in the painting.

Burnt_Umber

Logical. Whether the color "Burnt Umber" was used in the painting.

Cadmium_Yellow

Logical. Whether the color "Cadmium Yellow" was used in the painting.

Dark_Sienna

Logical. Whether the color "Dark Sienna" was used in the painting.

Indian_Red

Logical. Whether the color "Indian Red" was used in the painting.

Indian_Yellow

Logical. Whether the color "Indian Yellow" was used in the painting.

Liquid_Black

Logical. Whether the color "Liquid Black" was used in the painting.

Liquid_Clear

Logical. Whether the color "Liquid Clear" was used in the painting.

Midnight_Black

Logical. Whether the color "Midnight Black" was used in the painting.

Phthalo_Blue

Logical. Whether the color "Phthalo Blue" was used in the painting.

Phthalo_Green

Logical. Whether the color "Phthalo Green" was used in the painting.

Prussian_Blue

Logical. Whether the color "Prussian Blue" was used in the painting.

Sap_Green

Logical. Whether the color "Sap Green" was used in the painting.

Titanium_White

Logical. Whether the color "Titanium White" was used in the painting.

Van_Dyke_Brown

Logical. Whether the color "Van Dyke Brown" was used in the painting.

Yellow_Ochre

Logical. Whether the color "Yellow Ochre" was used in the painting.

Alizarin_Crimson

Logical. Whether the color "Alizarin Crimson" was used in the painting.

Source

Data collected from Bob Ross's The Joy of Painting, extracted from https://www.twoinchbrush.com/ and https://www.youtube.com/.